+++ pybillomat/errors.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+# coding: utf-8
+"""
+Errors
+"""
+
+
+class BillomatError(RuntimeError):
+    pass
+
+
+class EmptyFilterError(BillomatError):
+    pass
+